float64 to avoid overflow with certain counters

+++ src/collectors/zapi/collector/zapi.go
@@ -183,7 +183,7 @@ func (c *Zapi) PollInstance() (*matrix.Matrix, error) {
         }
     }
 
-    c.Metadata.SetValueSS("count", "instance", float32(count))
+    c.Metadata.SetValueSS("count", "instance", float64(count))
     logger.Info(c.Prefix, "added %d instances to cache (old cache had %d)", count, old_count)
 
     if len(c.Data.Instances) == 0 {
@@ -214,7 +214,7 @@ func (c *Zapi) PollData() (*matrix.Matrix, error) {
                     logger.Warn(c.Prefix, "%sSkipping metric [%s]: failed to parse [%s] float%s", util.Red, key, content, util.End)
                     skipped += 1
                 } else {
-                    c.Data.SetValue(metric, instance, float32(float))
+                    c.Data.SetValue(metric, instance, float64(float))
                     logger.Trace(c.Prefix, "%sMetric [%s] - Set Value [%f]%s", util.Green, key, float, util.End)
                     count += 1
                 }
