+++ src/collectors/zapi/plugins/shelf/shelf.go
@@ -216,7 +216,7 @@ func (p *Shelf) Run(data *matrix.Matrix) ([]*matrix.Matrix, error) {
 
 					if value := obj.GetChildContentS(mkey); value != "" {
 						if num, err := strconv.ParseFloat(value, 32); err == nil {
-							data.SetValue(m, instance, float32(num))
+							data.SetValue(m, instance, float64(num))
 							logger.Debug(p.Prefix, "Added numeric [%s] = [%f]", mkey, num)
 						} else {
 							logger.Warn(p.Prefix, "Failed to convert [%s] = [%s]", mkey, value)
